requirements.txt
langchain>=0.2.10
langchain-community>=0.2.9
langchain-chroma>=0.1.2
langchain-groq>=0.1.5
chromadb>=0.5.5
sentence-transformers>=3.0.1
pypdf>=4.3.1
python-dotenv>=1.0.1
streamlit>=1.37.1
uvicorn>=0.30.3
fastapi>=0.111.0
tiktoken>=0.7.0
numpy>=1.26.4
langchain-huggingface>=0.1.0
